Added the scripts
So yeah, this is what I call esx_impound, it has a dependency of es_extended if you have the job role of a Law Enforcement Officer, you will be elgible to use /impound in order to take the car to the impound. I'll make a Config soon so you can set where the Impound goes, it's not the most user friendly right now but it was made by me in a marketplace community full of developers who should understand how to edit basic co-ordinates more than most other things.

Any problems please let me know by emailing kyle@alv.gg or Alv#9999 on Discord, all feedback appreciated.

This script is also optimised, not that a script this small should have an affect on performance at all, but server total MS is becoming more popular lately.